refactor(tui): split renderer helpers
